update lpm upgrade  functions

+++ app/lpm/lua/update.lua
@@ -1,41 +1,40 @@
-local core      = require('core')
 local fs        = require('fs')
-local http      = require('http')
 local json      = require('json')
-local miniz     = require('miniz')
 local path      = require('path')
-local thread    = require('thread')
-local timer     = require('timer')
 local url       = require('url')
 local utils     = require('util')
-local qstring   = require('querystring')
 
 local request  	= require('http/request')
 local conf   	= require('app/conf')
 local ext   	= require('app/utils')
+local upgrade 	= require('./upgrade')
+
+-------------------------------------------------------------------------------
+
+-- Update states
+local STATE_INIT = 0
+local STATE_DOWNLOADING = 1
+local STATE_DOWNLOAD_COMPLETED = 2
+local STATE_UPDATING = 3
+
+-- Update result code
+local UPDATE_INIT = 0
+local UPDATE_SUCCESSFULLY = 1
+local UPDATE_NOT_ENOUGH_FLASH = 2
+local UPDATE_NOT_ENOUGH_RAM = 3
+local UPDATE_DISCONNECTED = 4
+local UPDATE_VALIDATION_FAILED = 5
+local UPDATE_UNSUPPORTED_FIRMWARE_TYPE = 6
+local UPDATE_INVALID_URI = 7
+local UPDATE_FAILED = 8
+local UPDATE_UNSUPPORTED_PROTOCOL = 9
+
+-------------------------------------------------------------------------------
 
 local exports = {}
 
-local formatFloat 		= ext.formatFloat
-local formatBytes 		= ext.formatBytes
-local noop 		  		= ext.noop
-local getSystemTarget 	= ext.getSystemTarget
-
 local function getNodePath()
 	return conf.rootPath
-end
-
-local function getRootPath()
-	return '/'
-end
-
-local function getRootURL()
-	local rootURL = require('app').rootURL
-	if (not rootURL) then
-		print('Unknown root URL.')
-	end
-
-	return rootURL
 end
 
 -------------------------------------------------------------------------------
@@ -57,15 +56,37 @@ local function parseVersion(version)
 		 + (parseVersionNumber(tokens[3]))
 end
 
--- Download system patch file
-local function downloadSystemPackage(options, callback)
-	callback = callback or noop
-
+local function saveUpdateStatus(status) 
 	local nodePath  = getNodePath()
 	local basePath  = path.join(nodePath, 'update')
 	local ret, err = fs.mkdirpSync(basePath)
 	if (err) then
 		print(err)
+		return
+	end
+
+	local filename = path.join(basePath, 'status.json')
+	local filedata = json.stringify(status);
+	fs.writeFileSync(filename, filedata);
+end
+
+--
+-- Download firmware file
+-- @param {object} options 
+--  - options.type
+--  - options.did
+--  - options.base
+--  - options.packageInfo
+-- @param {function} callback
+local function downloadFirmwarePackage(options, callback)
+	callback = callback or ext.noop
+	local printInfo = options.printInfo or function() end
+
+	local nodePath  = getNodePath()
+	local basePath  = path.join(nodePath, 'update')
+	local ret, err = fs.mkdirpSync(basePath)
+	if (err) then
+		callback({ code = UPDATE_NOT_ENOUGH_FLASH, error = err })
 		return
 	end
 
@@ -82,18 +103,35 @@ local function downloadSystemPackage(options, callback)
 			--print('md5sum', md5sum)
 
 			if (md5sum == packageInfo.md5sum) then
-				print("The update file is up-to-date!", filename)
+				-- print("The update file is up-to-date!", filename)
 				callback(nil, filename)
 				return
 			end
 		end
 	end
 
+	local rootURL 	= options.base
+	if (not rootURL) then
+		callback({ code = UPDATE_INVALID_URI, error = 'Invalid URI' })
+		return
+	end
+
+	local target 	= ext.getSystemTarget()
+	local version   = process.version or ''
+
+	--console.log(filename)
+	local baseURL 	= rootURL .. 'device/firmware/file'
+	local url 		= baseURL .. '?type=' .. (options.type or 'sdk')
+	url = url .. '&target=' .. target
+	url = url .. '&localVersion=' .. version
+	url = url .. '&did=' .. options.did
+
+	printInfo('Package url: ' .. url)
+
 	-- 下载最新的 SDK 更新包
-	request.download(options.url, {}, function(err, percent, response)
-		if (err) then 
-			print(err)
-			callback(err)
+	request.download(url, {}, function(err, percent, response)
+		if (err) then
+			callback({ code = UPDATE_DISCONNECTED, error = err })
 			return 
 		end
 
@@ -114,10 +152,17 @@ local function downloadSystemPackage(options, callback)
 		-- write to a temp file
 		print('Done!')
 
+		local filedata = response.body
+		local md5sum = utils.bin2hex(utils.md5(filedata))
+		if (md5sum ~= packageInfo.md5sum) then
+			callback({ code = UPDATE_VALIDATION_FAILED, error = 'Invalid firmware md5sum' })
+			return
+		end
+
 		os.remove(filename)
-		fs.writeFile(filename, response.body, function(err)
+		fs.writeFile(filename, filedata, function(err)
 			if (err) then 
-				callback(err)
+				callback({ code = UPDATE_NOT_ENOUGH_FLASH, error = err })
 				return
 			end
 
@@ -126,29 +171,35 @@ local function downloadSystemPackage(options, callback)
 	end)
 end
 
--- Download system 'package.json' file
-local function downloadSystemInfo(options, callback)
+-- 
+-- Download system 'update.json' file
+-- @param {object} options
+--  - options.type
+--  - options.did
+--  - options.base
+-- @param {function} callback
+local function downloadFirmwareInfo(options, callback)
 	options = options or {}
 	local printInfo = options.printInfo or function() end
 
 	-- URL
-	local arch      = os.arch()
-	local target 	= getSystemTarget()
-	local rootURL 	= getRootURL()
-	if (not rootURL) then
+	local did 		= options.did
+	local rootURL 	= options.base
+	if (not did) then
+		callback({ code = UPDATE_INVALID_URI, error = 'Invalid did' })
+		return
+
+	elseif (not rootURL) then
+		callback({ code = UPDATE_INVALID_URI, error = 'Invalid base URI' })
 		return
 	end
 
-	local baseURL 	= rootURL .. '/download/dist/' .. target
-	local url 		= baseURL .. '/nodelua-' .. target .. '-' .. (options.type or 'sdk') .. '.json'
-
-	printInfo("System target: " .. target)
-	printInfo("Upgrade server: " .. rootURL)	
+	local url 	= rootURL .. 'device/firmware/?did=' .. did
 	printInfo('URL: ' .. url)
 
 	request.download(url, {}, function(err, percent, response)
 		if (err) then
-			callback(err)
+			callback({ code = UPDATE_DISCONNECTED, error = err })
 			return
 
 		elseif (percent < 100) or (not response) then
@@ -158,34 +209,33 @@ local function downloadSystemInfo(options, callback)
 		--console.log(response.body)
 		local packageInfo = json.parse(response.body)
 		if (not packageInfo) or (not packageInfo.version) then
-			callback("Invalid system package information.")
+			callback({ code = UPDATE_VALIDATION_FAILED, error = "Invalid firmware package information." })
 			return
 		end
 
 		local version = parseVersion(packageInfo.version)
 		local current = parseVersion(process.version)
 		if (version < current) then
-			print('The update package version is less than current version: v' .. tostring(packageInfo.version))
-			return
+			print('Note: The firmware package version is less than current version: v' .. tostring(packageInfo.version) .. '<v' .. tostring(process.version))
 		end
 
 		local nodePath  = getNodePath()
 		local basePath  = path.join(nodePath, 'update')
 		local ret, err = fs.mkdirpSync(basePath)
 		if (err) then
-			print(err)
+			callback({ code = UPDATE_NOT_ENOUGH_FLASH, error = err })
 			return
 		end
 
-		local filename 	= path.join(basePath, 'package.json')
+		local filename 	= path.join(basePath, 'update.json')
 		local filedata  = fs.readFileSync(filename)
 		if (filedata == response.body) then
-			print("The system information is up-to-date!")
+			-- print("The firmware information is up-to-date!")
 			callback(nil, packageInfo)
 			return
 		end
 
-		local tempname 	= path.join(basePath, 'package.json.tmp')
+		local tempname 	= path.join(basePath, 'update.json.tmp')
 		os.remove(tempname)
 
 		fs.writeFile(tempname, response.body, function(err)
@@ -197,91 +247,117 @@ local function downloadSystemInfo(options, callback)
 			os.remove(filename)
 			local ret, err = os.rename(tempname, filename)
 			if (err) then
-				print(err)
+				callback({ code = UPDATE_NOT_ENOUGH_FLASH, error = err })
 				return
 			end
 
-			print("System information saved to: " ..  filename)
+			print("Firmware information saved to: " ..  filename)
 			callback(nil, packageInfo)
 		end)
 	end)
 end
 
--- Download system 'package.json' and patch files
-local function downloadUpdateFiles(options, callback)
+--
+-- Download 'update.json' and firmware files
+-- @param {object} options
+--  - options.type
+--  - options.did
+-- @param {function} callback
+local function downloadFirmwareFile(options, callback)
 	options = options or {}
 	local printInfo = options.printInfo or function() end
-
-	-- System update URL
-	local target 	= getSystemTarget()
-	local rootURL 	= getRootURL() or 'http://127.0.0.1'
-	if (not rootURL) then
-		return
-	end
 
 	-- Update Path
 	local nodePath  = getNodePath()
 	local basePath  = path.join(nodePath, 'update')
 	local ret, err = fs.mkdirpSync(basePath)
 	if (err) then
-		print(err)
+		callback({ code = UPDATE_NOT_ENOUGH_FLASH, error = err })
 		return
 	end
 
-	local md5sum      = ''
-	local version   = process.version or ''
+	-- downloading firmware info
+	downloadFirmwareInfo(options, function(err, packageInfo)
+		-- console.log('Firmware Info:', packageInfo);
+		if (err) or (not packageInfo) then
+			callback(err)
+			return
+		end
 
-	-- package.json
-	local filename 	= path.join(basePath, 'package.json')
-	local filedata  = fs.readFileSync(filename)
-	local packageInfo = json.parse(filedata)
-	if (packageInfo) then
-		md5sum = packageInfo.md5sum or ''
-	end
+		-- downloading firmware file
+		local args = {}
+		args.did 		 = options.did
+		args.printInfo   = options.printInfo
+		args.type        = options.type
+		args.packageInfo = packageInfo
+		downloadFirmwarePackage(args, function(err, filename)
+			if (err) or (not filename) then
+				return callback(err)
+			end
 
-	--console.log(filename)
+			-- check firmware file format
+			local rootPath = os.tmpdir .. '/update'
 
-	local baseURL 	= rootURL .. '/update/'
-	local url 		= baseURL .. '&type=' .. (options.type or 'sdk') 
-	url 		= url .. '?target=' .. target
-	url 		= url .. '&md5sum=' .. md5sum
-	url 		= url .. '&version=' .. version
+			local options = {
+				filename = filename,
+				rootPath = rootPath
+			}
+			local updater = upgrade.openUpdater(options)
+			updater.reader = upgrade.openBundle(filename)
 
-	printInfo('Package url: ' .. url)
+			local packageInfo = updater:parsePackageInfo()
+			local version = packageInfo and packageInfo.version
+			if (not version) then
+				console.log('packageInfo', packageInfo)
+				return callback({ code = UPDATE_VALIDATION_FAILED, error = 'Invalid firmware file format'});
+			end
 
-	-- downloading
-	local args = {}
-	args.url 		 = url
-	args.type        = options.type
-	args.packageInfo = packageInfo
-	downloadSystemPackage(args, function(err, filename)
-		printInfo("Done.")
-		callback(err, filename, packageInfo)
+			--local newname = basePath .. '/update-' .. version .. '.zip'
+			--os.remove(newname)
+			--os.rename(filename, newname)
+			callback(nil, filename, packageInfo)
+		end)
 	end)
 end
 
+-------------------------------------------------------------------------------
+--
 
 -- Dnowload update files only
 function exports.update(callback)
-	local options = { type = 'patch' }
+	local options = { type = 'update' }
+
+	local profile = conf('user')
+	options.did = profile:get('did')
+	options.base = profile:get('base')
 	
 	if (type(callback) ~= 'function') then
 		callback = function(err, filename, packageInfo)
 			packageInfo = packageInfo or {}
+			local status = { state = 0, result = 0 }
 
 			--console.log(err, filename, packageInfo)
 			if (err) then
-				print('err: ', err)
+				console.log('err: ', err)
+				status.result = 8
+				if (err.result) then
+					status.result = err.result
+				end
 
 			else
-				print('latest version: ' .. tostring(packageInfo.version))
+				local version = packageInfo.version
+				print('Latest firmware version: ' .. tostring(version))
+				status.state = STATE_DOWNLOAD_COMPLETED
+				status.version = version
 			end
+
+			saveUpdateStatus(status)
 		end
 
 		options.printInfo = function(...) print(...) end
 	end
-
-	downloadUpdateFiles(options, callback)
+	
+	downloadFirmwareFile(options, callback)
 end
 
 return exports
